+++ changelog.md
@@ -1,4 +1,7 @@
 ## [1.0.0-beta6.1] - 2026-04-28
 
+### Added
+- Colored Vault UIs: Vaults from *Create: Vibrant Vaults* now have a UI that matches their block color (can be disabled in config).
+
 ### Changed
 - Refined the shift-double-click "Take All" mechanic to only trigger when an item is already held in the cursor. This prevents accidental bulk transfers when trying to rapidly extract individual stacks.